+++ app/src/main/java/org/simple/clinic/summary/medicalhistory/MedicalHistorySummaryUi.kt
@@ -2,7 +2,7 @@ package org.simple.clinic.summary.medicalhistory
 
 import org.simple.clinic.medicalhistory.MedicalHistory
 
-interface MedicalHistorySummaryUi: MedicalHistorySummaryUiActions {
+interface MedicalHistorySummaryUi {
   fun populateMedicalHistory(medicalHistory: MedicalHistory)
   fun showDiagnosisView()
   fun hideDiagnosisView()