+++ app/src/main/java/org/simple/clinic/summary/medicalhistory/MedicalHistorySummaryUpdate.kt
@@ -4,13 +4,9 @@ import com.spotify.mobius.Next
 import com.spotify.mobius.Next.dispatch
 import com.spotify.mobius.Update
 import org.simple.clinic.medicalhistory.MedicalHistory
-import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.DIAGNOSED_WITH_DIABETES
-import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.DIAGNOSED_WITH_HYPERTENSION
 import org.simple.clinic.mobius.next
 
 class MedicalHistorySummaryUpdate : Update<MedicalHistorySummaryModel, MedicalHistorySummaryEvent, MedicalHistorySummaryEffect> {
-
-  private val diagnosisQuestions = setOf(DIAGNOSED_WITH_HYPERTENSION, DIAGNOSED_WITH_DIABETES)
 
   override fun update(
       model: MedicalHistorySummaryModel,
@@ -30,9 +26,6 @@ class MedicalHistorySummaryUpdate : Update<MedicalHistorySummaryModel, MedicalHi
     val effects = mutableSetOf<MedicalHistorySummaryEffect>()
 
     val toggledQuestion = event.question
-    if (toggledQuestion in diagnosisQuestions) {
-      effects.add(HideDiagnosisError)
-    }
 
     val updatedMedicalHistory = savedMedicalHistory.answered(toggledQuestion, event.answer)
     effects.add(SaveUpdatedMedicalHistory(updatedMedicalHistory))