+++ app/src/main/java/org/simple/clinic/summary/medicalhistory/MedicalHistorySummaryView.kt
@@ -5,6 +5,7 @@ import android.os.Parcelable
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.widget.FrameLayout
+import androidx.appcompat.app.AppCompatActivity
 import io.reactivex.Observable
 import io.reactivex.rxkotlin.ofType
 import io.reactivex.subjects.PublishSubject
@@ -20,6 +21,7 @@ import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.DIAGNOSED_WITH_HY
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_HEART_ATTACK
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_KIDNEY_DISEASE
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HAS_HAD_A_STROKE
+import org.simple.clinic.medicalhistory.SelectDiagnosisErrorDialog
 import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.router.screen.ScreenRouter
 import org.simple.clinic.summary.PatientSummaryChildView
@@ -41,7 +43,10 @@ class MedicalHistorySummaryView(
   lateinit var screenRouter: ScreenRouter
 
   @Inject
-  lateinit var effectHandlerFactory: MedicalHistorySummaryEffectHandler.Factory
+  lateinit var activity: AppCompatActivity
+
+  @Inject
+  lateinit var effectHandler: MedicalHistorySummaryEffectHandler
 
   init {
     LayoutInflater.from(context).inflate(R.layout.medicalhistory_summary_view, this, true)
@@ -68,7 +73,7 @@ class MedicalHistorySummaryView(
         defaultModel = MedicalHistorySummaryModel.create(screenKey.patientUuid),
         update = MedicalHistorySummaryUpdate(),
         init = MedicalHistorySummaryInit(),
-        effectHandler = effectHandlerFactory.create(this).build(),
+        effectHandler = effectHandler.build(),
         modelUpdateListener = { model ->
           modelUpdateCallback?.invoke(model)
           uiRenderer.render(model)
@@ -142,16 +147,12 @@ class MedicalHistorySummaryView(
     kidneyDiseaseQuestionView.hideDivider()
   }
 
-  override fun hideDiagnosisError() {
-    diagnosisRequiredError.visibility = GONE
-  }
-
   override fun registerSummaryModelUpdateCallback(callback: PatientSummaryModelUpdateCallback?) {
     modelUpdateCallback = callback
   }
 
   fun showDiagnosisError() {
-    diagnosisRequiredError.visibility = VISIBLE
+    SelectDiagnosisErrorDialog.show(activity.supportFragmentManager)
   }
 
   private fun answerToggled(question: MedicalHistoryQuestion, answer: Answer) {